+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3890,7 +3890,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M2G6X6YWQP46VWCVR13GCASC"),
+  id: z.string().default("01M2HQ2DNQEANMKDEWJ3PBPY1D"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),