+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-14T14:59:49.769Z",
+    "2026-09-15T05:01:32.178Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {